Moved Strelka tests out of the installed module.
The tests shouldn't really be installed as part of the module, and moving them outside
the tree makes it conceptually clearer that they are -not- part of the strelka module at
all (since that's how pytest works). As a compromise, the previous common code in
strelka.tests has been moved to strelka.testing.

Removed extraneous scanner import test, since, with the scanners as fixtures, this is
automatically tested whenever a scanner is tested.